Restructure into route/config
