+++ client/src/route/config/ConfigItems.tsx
@@ -2,7 +2,7 @@ import CheckCircleIcon from '@mui/icons-material/CheckCircle';
 import ErrorOutlineIcon from '@mui/icons-material/ErrorOutline';
 import { Tooltip } from '@mui/material';
 import * as React from 'react';
-import { validationType } from 'util/config';
+import { validationType } from 'route/config/Verify';
 import { StatusCodes } from 'http-status-codes';
 
 const ConfigIcon = (toolTipTitle: string, icon: JSX.Element): JSX.Element => (
@@ -48,19 +48,19 @@ export const ConfigItem: React.FC<{
     value: string;
     validation?: validationType;
 }> = ({ label, value, validation = { error: 'Validation unavailable' } }) => (
-        <div
-            style={{
-                display: 'flex',
-                alignItems: 'center',
-                gap: '10px',
-                margin: '5px 0',
-            }}
-            className="Config-item"
-        >
-            {getConfigIcon(validation, label)}
-            <div id="config-text">
-                <strong>{label}:</strong> {value}
-            </div>
+    <div
+        style={{
+            display: 'flex',
+            alignItems: 'center',
+            gap: '10px',
+            margin: '5px 0',
+        }}
+        className="Config-item"
+    >
+        {getConfigIcon(validation, label)}
+        <div id="config-text">
+            <strong>{label}:</strong> {value}
         </div>
-    );
+    </div>
+);
 ConfigItem.displayName = 'ConfigItem';
