+++ client/src/routes.tsx
@@ -3,7 +3,7 @@ import WorkBench from 'route/workbench/Workbench';
 import LayoutPublic from 'page/LayoutPublic';
 import PrivateRoute from 'route/auth/PrivateRoute';
 import LibraryPreview from 'preview/route/library/LibraryPreview';
-import VerifyConfig from 'util/config';
+import VerifyConfig from 'route/config/Verify';
 import Library from './route/library/Library';
 import DigitalTwins from './route/digitaltwins/DigitalTwins';
 import DigitalTwinsPreview from './preview/route/digitaltwins/DigitalTwinsPreview';
@@ -11,70 +11,70 @@ import SignIn from './route/auth/Signin';
 import Account from './route/auth/Account';
 
 export const routes = [
-  {
-    path: '/',
-    element: (
-      <LayoutPublic>
-        <SignIn />
-      </LayoutPublic>
-    ),
-  },
-  {
-    path: 'verify',
-    element: (
-      <LayoutPublic containerMaxWidth="md">
-        <VerifyConfig />
-      </LayoutPublic>
-    ),
-  },
-  {
-    path: 'library',
-    element: (
-      <PrivateRoute>
-        <Library />
-      </PrivateRoute>
-    ),
-  },
-  {
-    path: 'digitaltwins',
-    element: (
-      <PrivateRoute>
-        <DigitalTwins />
-      </PrivateRoute>
-    ),
-  },
-  {
-    path: 'account',
-    element: (
-      <PrivateRoute>
-        <Account />
-      </PrivateRoute>
-    ),
-  },
-  {
-    path: 'workbench',
-    element: (
-      <PrivateRoute>
-        <WorkBench />
-      </PrivateRoute>
-    ),
-  },
-  {
-    path: 'preview/digitaltwins',
-    element: (
-      <PrivateRoute>
-        <DigitalTwinsPreview />
-      </PrivateRoute>
-    ),
-  },
-  {
-    path: 'preview/library',
-    element: (
-      <PrivateRoute>
-        <LibraryPreview />
-      </PrivateRoute>
-    ),
-  },
+    {
+        path: '/',
+        element: (
+            <LayoutPublic>
+                <SignIn />
+            </LayoutPublic>
+        ),
+    },
+    {
+        path: 'verify',
+        element: (
+            <LayoutPublic containerMaxWidth="md">
+                <VerifyConfig />
+            </LayoutPublic>
+        ),
+    },
+    {
+        path: 'library',
+        element: (
+            <PrivateRoute>
+                <Library />
+            </PrivateRoute>
+        ),
+    },
+    {
+        path: 'digitaltwins',
+        element: (
+            <PrivateRoute>
+                <DigitalTwins />
+            </PrivateRoute>
+        ),
+    },
+    {
+        path: 'account',
+        element: (
+            <PrivateRoute>
+                <Account />
+            </PrivateRoute>
+        ),
+    },
+    {
+        path: 'workbench',
+        element: (
+            <PrivateRoute>
+                <WorkBench />
+            </PrivateRoute>
+        ),
+    },
+    {
+        path: 'preview/digitaltwins',
+        element: (
+            <PrivateRoute>
+                <DigitalTwinsPreview />
+            </PrivateRoute>
+        ),
+    },
+    {
+        path: 'preview/library',
+        element: (
+            <PrivateRoute>
+                <LibraryPreview />
+            </PrivateRoute>
+        ),
+    },
 ];
 
 export default routes;
