+++ client/test/unit/routes/SignIn.test.tsx
@@ -9,7 +9,7 @@ import {
 import { MemoryRouter } from 'react-router-dom';
 import SignIn from 'route/auth/Signin';
 import { useAuth } from 'react-oidc-context';
-import { getValidationResults } from 'util/config'; // Globally mocked
+import { getValidationResults } from 'route/config/Verify'; // Globally mocked
 
 jest.unmock('route/auth/Signin');
 jest.mock('react-oidc-context');
@@ -94,7 +94,7 @@ describe('SignIn', () => {
         });
 
         await waitFor(() => {
-            expect(screen.getByText(/Config validation failed/i)).toBeInTheDocument();
+            expect(screen.getByText(/Invalid Application Configuration. Please contact the administrator of your DTaaS installation./i)).toBeInTheDocument();
         });
     });
